pyproject.toml
[tool.poetry]
name = "portfolio_api"
version = "0.1.0"
description = ""
authors = ["Christian Gonzalez <christiangonzalezblack@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.6"
psycopg2-binary = "^2.8.6"
whitenoise = "^5.2.0"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.2"
django-allauth = "^0.44.0"
dj-rest-auth = "^2.1.3"
django-cors-headers = "^3.7.0"
Pillow = "^8.1.0"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
